moved token endpoints to user_endpoint.py
